// src/lib/server/types/discord/oauth.ts
export type OAuth = {
  id: string;
  username: string;
  avatar: string;
  discriminator: string;
  public_flags: number;
  flags: number;
  banner: null;
  accent_color: number;
  global_name: null;
  avatar_decoration_data: null;
  banner_color: string;
  clan: null;
  mfa_enabled: boolean;
  locale: string;
  premium_type: number;
};